server/requirements.txt
google-genai
google-adk
google-cloud-storage
google-cloud-firestore
python-dotenv
fastapi
uvicorn
google-generativeai
Pillow